backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pinecone-client>=2.2.4
cohere>=4.37
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3